# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/products": {
      "get": {
        "summary": "Get Products",
        "operationId": "get_products_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "items": { "$ref": "#/components/schemas/Product" }, "type": "array", "title": "Response Get Products Products Get" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Product",
        "operationId": "create_product_products_post",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }, "required": true },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "items": { "$ref": "#/components/schemas/Product" }, "type": "array", "title": "Response Create Product Products Post" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "summary": "Get Product",
        "operationId": "get_product_products__product_id__get",
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Product Id" } }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" },
                  "title": "Response Get Product Products  Product Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "put": {
        "summary": "Update Product",
        "operationId": "update_product_products__product_id__put",
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Product Id" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "summary": "Delete Product",
        "operationId": "delete_product_products__product_id__delete",
        "parameters": [{ "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Product Id" } }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "string" },
                  "title": "Response Delete Product Products  Product Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Message",
        "operationId": "message__get",
        "responses": { "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } } }
      }
    },
    "/goodbye": {
      "get": {
        "summary": "Goodbye",
        "operationId": "goodbye_goodbye_get",
        "responses": { "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } } }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": { "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" } },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Product": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "price": { "type": "number", "title": "Price" },
          "stock": { "type": "integer", "title": "Stock" }
        },
        "type": "object",
        "required": ["id", "name", "price", "stock"],
        "title": "Product"
      },
      "ValidationError": {
        "properties": {
          "loc": { "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }, "type": "array", "title": "Location" },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}